I made a Asteroid Spawner
It's not finished, right now it's a mess,I have to implement factory.

This spawner should be able to:
spawn the asteroids in a random point on the right side of the screen
it has to be able to increment the speed it spawns the asteroids and the amount of them over time.
And either asign the random values of the asteroids or instead of the spawn asigning those values, it is the asteroid that asign those values to itself.
if it spawns more than one asteroid it needs to check if the asteroids didn't spawned in the same position.