I made the final boss prefab, but it does nothing yet
To kill this bos you have to destroy all his weapons, The healthbar should display the life of all the weapons. It will have 2 states:
The first one it has 2 weapons with a bigger hitbox that shoots in two directions and rotate.
The second one has 4 guns, with a smoller hitbox that shoot straight and the boss rotates