Progress with the asteroids
The asteroids spawn with a random size between 2 values, a random rotation, random direction.
Next:
Their health and damage should be multiplied by the size (That is so bigger asteroids inflict more damge and are harder to kill than the little ones).

There should something that detects if a asteroid spawned inside another and give it another position.

The asteroids should spawn 2 or 3 asteroids when destroyed.

Strore the random values in a scriptableobject.