I found a bug in sine movement and I modified it
For some reason when you modified the frequency it made the sine movement amplitude bigger (not the amplitude of the scriptable object, but from the sine wave), so I change the code a little and now it doesn't do it any more.

The previous code is commented if you want to check it out.